+++ DBServer/Models/Shift.cs
@@ -3,7 +3,7 @@ using System;
 namespace DBServer.Models {
     public class Shift {
         public Shift () { }
-        public Shift (long v1, long v2, DateTime dateTime, int v3, string v4) {
+        public Shift (long v1, long v2, DateTime dateTime, double v3, string v4) {
             this.ShiftId = v1;
             this.CompanyId = v2;
             this.ShiftDate = dateTime;
@@ -13,7 +13,7 @@ namespace DBServer.Models {
         public long ShiftId { get; set; }
         public long CompanyId { get; set; }
         public DateTime ShiftDate { get; set; }
-        public int ShiftTime { get; set; }
+        public double ShiftTime { get; set; }
         public string ShiftRequirements { get; set; }
